ByteData takes a long -- no longer confined to a 2GB FastBlob, or 2GB of data for a single type in ByteArrayOrdinalMap

+++ src/main/java/com/netflix/zeno/fastblob/record/ByteDataBuffer.java
@@ -27,7 +27,7 @@ package com.netflix.zeno.fastblob.record;
 public class ByteDataBuffer {
 
     private final SegmentedByteArray buf;
-    private int position;
+    private long position;
 
     public ByteDataBuffer() {
         this(256);
@@ -46,20 +46,20 @@ public class ByteDataBuffer {
         position = 0;
     }
 
-    public void setPosition(int position) {
+    public void setPosition(long position) {
         this.position = position;
     }
 
-    public int length() {
+    public long length() {
         return position;
     }
 
     public void copyTo(ByteDataBuffer other) {
-        other.buf.copy(buf, 0, other.position, position);
+        other.buf.copy(buf, 0, other.position, (int)position);
         other.position += position;
     }
 
-    public byte get(int index) {
+    public byte get(long index) {
         return buf.get(index);
     }
